🎆 Separated Style Logic From Actual Page Components, TODO : Refactor rest of pages, componennts
