+++ src/pages/Normatividad/index.jsx
@@ -1,0 +1,2 @@
+import Normatividad from "./Normatividad";
+export default Normatividad;
