+++ src/pages/Solicitudes/index.jsx
@@ -1,0 +1,2 @@
+import Solicitudes from "./Solicitudes";
+export default Solicitudes;
